+++ platform/viewer/src/setupTools.js
@@ -142,8 +142,8 @@ function getResetLabellingAndContextMenu(store) {
 export default function setupTools(store) {
   const toolLabellingFlowCallback = getToolLabellingFlowCallback(store);
   const availableTools = [
-    { name: 'Pan', mouseButtonMasks: [4] },
-    { name: 'Zoom', mouseButtonMasks: [2] },
+    { name: 'Pan', mouseButtonMasks: [1, 4] },
+    { name: 'Zoom', mouseButtonMasks: [1, 2] },
     { name: 'Wwwc', mouseButtonMasks: [1] },
     { name: 'Magnify' },
     { name: 'WwwcRegion' },